add init file to pages package
add init file and move Routine enum class definition to init file
