+++ pages/__init__.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+
+class Routine(Enum):
+    SLEEP = "Sleeping Time"
+    CLASS = "Class Time"
+    MEAL = "Meal Time"
+    STUDY = "Study Time"
+    EXERCISE = "Exercise Time"
