+++ pages/comparison.py
@@ -1,10 +1,11 @@
-from enum import Enum
+
 
 import dash
 import dash_daq as daq
 from dash import Dash, html, dcc, callback
 from dash.dependencies import Input, Output
 
+from pages import Routine
 from figures.location_mapbox import location_mapbox, location_mapbox_fake_data
 from figures.weekly_routine_timeline import weekly_routine_timeline, weekly_routine_fake_data
 from figures import BFI, daily_routine, indoor
@@ -18,16 +19,6 @@ df_user, df_roommate = fake_data.user_and_roommate_data()
 fig_bfi = BFI.bfi_fig(df_user[0], df_roommate[0]).fig_cmp
 fig_indoor = indoor.indoor_fig(df_user[2], df_roommate[2]).fig_cmp
 fig_daily_routine = daily_routine.daily_routine_fig(df_user[1], df_roommate[1]).fig_cmp
-
-
-
-
-class Routine(Enum):
-    SLEEP = "Sleeping Time"
-    CLASS = "Class Time"
-    MEAL = "Meal Time"
-    STUDY = "Study Time"
-    EXERCISE = "Exercise Time"
 
 
 layout = html.Div(children=[
